+++ src/vivarium/framework/engine.py
@@ -194,9 +194,7 @@ class SimulationContext:
                 "that do not inherit from `vivarium.Component`: "
                 f"[{[c.name for c in non_components]}]."
             )
-            self._logger.warning(message)
-            # TODO: raise error as part of MIC-4487
-            # raise ComponentConfigError(message)
+            raise ComponentConfigError(message)
 
         self._lifecycle.add_constraint(self.add_components, allow_during=["initialization"])
         self._lifecycle.add_constraint(