+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+**2.2.0 - 10/19/23**
+
+ - Refactor Manager configuration defaults
+ - Throw an error if simulation attempts to use a component that is not an instance of Component
+ - Refactor and simplify LookupTable implementations
+ - Enable LookupTable to have source data that is only categorical
+ - Enable LookupTables with DataFrame source data to specify value columns
+
 **2.1.1 - 10/13/23**
 
  - Enable RandomnessStream to sample from a distribution